Initial commit.
This is going to be an implementation of previous 3284 custom swerve
ported to Java. I can only hope I'm using WPILib idiomatically, since
that's generally the best way to avoid UB, but that also seems to be
less of an issue with Java than C++.

In any case, this code is mostly for example purposes only and does not
represent any actual robot, nor does it implement any of the needed
funcitons of a competition robot (Autonous control is Not Planned).

SPDX-License-Identifier: CC-BY-NC-SA-4.0
Signed-off-by: Ayzee Patton <vlpatton@vlpatton.gay>