Added pose estimation and vision measurement API to drivetrain class.
This will allow more integration with an actual vision system, which
would be good for teaching and demonstration vision pose estimation.
It's only theory tho, since vision is not a goal of this codebase.
Modifications or derivations of it may contain such implementations of
vision, and that would be a good exercise for the reader :)

In any case, the next step is the actual loop for when the drivetrain is
enabled. Autonomous control will not really be implemented until future
revisions or derivations of this codebase, so a tag should be coming
soon.

SPDX-License-Identifier: CC-BY-NC-SA-4.0
Signed-off-by: Ayzee Patton <vlpatton@vlpatton.gay>